+++ docs/assets/cta-button.stories-DTmExRqz.js
@@ -1,0 +1,22 @@
+import{t as y}from"./index-CCz6reEH.js";import{t as _,r as g,b as f,e as h,p as k,f as B,d as x}from"./registry-B-klnak9.js";import{_ as C}from"./dynamic-tag-DQCvkDdb.js";import{_ as S}from"./index-BRJKYh5S.js";import{_ as w}from"./render-tag-BBOJ9dEX.js";import"./_commonjsHelpers-Cpj98o6Y.js";import"./_commonjs-dynamic-modules-TDtrdbi3.js";import"./index-C98gAawF.js";/* empty css             */const T=`<h1 style='display: flex; justify-content: space-between; align-items: center;'>
+    <span>
+        ebay-cta-button
+    </span>
+    <span style='font-weight: normal; font-size: medium; margin-bottom: -15px;'>
+        DS v1.1.0
+    </span>
+</h1>
+
+## Examples and Documentation
+
+-   [Storybook](https://ebay.github.io/ebayui-core/?path=/story/buttons-ebay-cta-button)
+-   [Storybook Docs](https://ebay.github.io/ebayui-core/?path=/docs/buttons-ebay-cta-button)
+-   [Code Examples](https://github.com/eBay/ebayui-core/tree/master/src/components/ebay-cta-button/examples)
+`,r="$IoRPYei",a=_(r),z=a;g.r(r,()=>a);const c={};a._=f(function(n,t,s,o,E,A){const{size:i,class:d,renderBody:u,...b}=n;t.be("a",h(k(b),{class:B(["cta-btn",`cta-btn--${i||"default"}`,d])}),"0",o,null,4),t.be("span",{class:"cta-btn__cell"},"1",o,null,1),t.be("span",null,"2",o,null,0),C(t,u,null,null,null,null,s,"3"),t.ee(),w(S,{},t,s,"4"),t.ee(),t.ee()},{t:r,i:!0},c);a.Component=x(c,a._);const $=n=>({input:{...n,renderBody(t){t.html(n.renderBody)}}}),Y={title:"buttons/ebay-cta-button",component:z,parameters:{docs:{description:{component:T}}},argTypes:{renderBody:{},size:{type:"options",description:'Can be "large"',table:{defaultValue:{summary:"default"}},options:["default","large"]},href:{description:"link target",table:{defaultValue:{summary:""}}}}},e=$.bind({});e.args={renderBody:"CTA button",href:"http://www.ebay.com",size:"regular"};e.parameters={docs:{source:{code:y("ebay-cta-button",e.args)}}};var m,l,p;e.parameters={...e.parameters,docs:{...(m=e.parameters)==null?void 0:m.docs,source:{originalSource:`args => ({
+  input: {
+    ...args,
+    renderBody(out) {
+      out.html(args.renderBody);
+    }
+  }
+})`,...(p=(l=e.parameters)==null?void 0:l.docs)==null?void 0:p.source}}};const q=["Standard"];export{e as Standard,q as __namedExportsOrder,Y as default};