+++ src/components/ebay-icon/icons/ebay-target-16-icon/symbol.ts
@@ -1,0 +1,4 @@
+export function symbol() {
+    // eslint-disable-next-line max-len,quotes
+    return '<symbol viewbox="0 0 16 16" id=icon-target-16><path fill-rule=evenodd clip-rule=evenodd d="M12 0a1 1 0 0 1 1 1v2h2a1 1 0 1 1 0 2h-2.586L9.328 8.086a2.5 2.5 0 1 1-1.414-1.414L11 3.586V1a1 1 0 0 1 1-1ZM2 9a5 5 0 0 1 5.835-4.93 1 1 0 0 0 .33-1.973 7 7 0 1 0 5.738 5.738 1 1 0 0 0-1.972.33A5 5 0 1 1 2 9Z"></path></symbol>';
+}