+++ src/components/ebay-phone-input/style.ts
@@ -1,2 +1,1 @@
 import "@ebay/skin/phone-input";
-import "@ebay/skin/flag";